requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jsonschema==4.23.0

# compatible with faiss-cpu 1.8.x
numpy==1.26.4
faiss-cpu==1.8.0.post1

# auth + security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# db
SQLAlchemy==2.0.36

# tools/tests
requests==2.32.5
pip-audit==2.7.3
